fix(build): move pure functions out of route.ts to fix Next.js export error
Next.js rejects non-HTTP-verb exports from API route files at build time:
  "isPDF" is not a valid Route export field.

Moved isPDF, parseProfileResponse, calculateCost and their shared types
(CMFItem, ExtractionConfidence, ExtractionResult) into a co-located
utils.ts file. route.ts imports from utils.ts; route.test.ts now also
imports from utils.ts directly.
